refactor: manager by uv tool Ⅱ
