+++ .github/workflows/pypi.yml
@@ -1,0 +1,31 @@
+name: Publish
+permissions:
+  contents: write
+  pull-requests: read
+on:
+  push:
+    tags:
+      - 'v*.*.*'  
+
+jobs:
+  publish:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v4
+      with:
+        fetch-depth: 1
+
+    - name: Install uv
+      uses: astral-sh/setup-uv@v4
+      with:
+        enable-cache: true
+
+    - name: Build package
+      run: uv build
+
+    - name: Publish to PyPI
+      env:
+        UV_PUBLISH_TOKEN: ${{ secrets.PYPI_TOKEN }}
+      run: uv publish
+
