+++ README.md
@@ -55,11 +55,26 @@
 
 #### 🎁 Features
 
+✅ 选网站可直接选kemono，单独窗口供后续脚本集扩展，详情看[📒使用指引](https://jasoneri.github.io/ComicGUISpider/feat/script)  
+✅ 自定义背景，[配置key](https://jasoneri.github.io/ComicGUISpider/config/#bg-path)  
+
+#### 🐞 Fix
+
+✅ 修复包缺失导致的无法运行、kb缓存时间逻辑、内部重启的 division by zero 报错等
+
+<details>
+<summary> <code>v2.4.0-beta</code> 开发版特性👈看就点</summary>
+
+#### 🎁 Features
+
 ✅ 已打包上传至 pypi ，可使用 uv tool 管理/运行 CGS [查看细则](https://jasoneri.github.io/ComicGUISpider/deploy/quick-start)  
 
 #### 🐞 Fix
 
 ✅ kemono域名更换  
+
+🍗使用开发版遇到报错时请积极提 issue 等提供线索，谢谢
+</details>
 
 > 配置窗口左下设有`检查更新`按钮，请根据提示进行更新操作
