+++ docs/faq/index.md
@@ -29,41 +29,22 @@
 
 ### ModuleNotFoundError: No module named 'xxx'
 
-win:
-
-1. 在绿色包解压的目录打开 (powershell) 终端执行命令  
-
-``` bash
-irm https://gitee.com/json_eri/ComicGUISpider/raw/GUI/deploy/online_scripts/win.ps1 | iex
-```
-
-::: info 非绿色包的用户参考 [🚀 快速开始 > 1-部署](../deploy/quick-start#1-下载--部署) 的安装依赖命令示例
-:::
-
-macOS: （与 mac部署 初始化步骤命令一样的）  
-
-```bash
-bash /Applications/CGS.app/Contents/Resources/scripts/deploy/launcher/mac/init.bash
-```
+截图或带上 `log` 提交 [issue](
+  https://github.com/jasoneri/ComicGUISpider/issues/new?template=bug-report.yml
+)
 
 ### 更新失败后程序无法打开
 
-::: tip 最简单有效❗️  
-下载📦绿色包 覆盖更新  
+::: tip 两种方式可选  
+
+- 下载📦绿色包 重新安装/更新  
+- 干脆直接使用 [uv tool方式部署安装](/deploy/quick-start#1-下载--部署)
 :::
-
-更新的报错日志已整合进 log/GUI.log 文件里，建议提 issue 并附上 log，帮助 CGS 进行优化  
-
-1. 若 CGS 还能打开，删除 `scripts/deploy/version.json`，重新用内置更新
-
-2. 不能打开时: 将最新版本的 `Source code (zip)` 源码包，解压后将全部源码覆盖到 scripts 目录下  
-
-2.1 按上面 ModuleNotFoundError 的方法安装依赖
 
 ### 【win】弹出消息框报错而且一堆英文不是中文(非开发者预设报错)的时候
 
 例如`Qxxx:xxxx`, `OpenGL`等，此前已优化过，如还有弹出框警告，  
-尝试在解压目录使用cmd运行`./CGS.bat > CGS-err.log 2>&1`，然后把`CGS-err.log`发群里反馈
+尝试在解压目录使终端执行`./CGS.exe > CGS-err.log 2>&1`，然后把`CGS-err.log`发群里反馈
 
 ---
